Update README.md
The Todo Cubit application aims to provide users with a seamless experience in task management. Built using Flutter, it leverages the power of the Cubit state management solution, allowing for a reactive and efficient interface. The app is designed to be lightweight, ensuring that it runs smoothly on a variety of devices while providing a clean and modern UI.

Whether you're a developer looking to learn about Flutter and state management or a user seeking a simple yet effective way to manage tasks, Todo Cubit is an excellent choice. It demonstrates the capabilities of Flutter and the Cubit package, making it a great starting point for anyone interested in modern app development.